Lift and shift into a component
