+++ app/src/ui/app.tsx
@@ -12,22 +12,19 @@ import { AppStore, GitHubUserStore, IssuesStore } from '../lib/stores'
 import { assertNever } from '../lib/fatal-error'
 import { shell } from '../lib/app-shell'
 import { updateStore, UpdateStatus } from './lib/update-store'
-import { RetryAction } from '../models/retry-actions'
 import { FetchType } from '../models/fetch'
 import { shouldRenderApplicationMenu } from './lib/features'
 import { matchExistingRepository } from '../lib/repository-matching'
 import { getDotComAPIEndpoint } from '../lib/api'
-import { getVersion, getName } from './lib/app-proxy'
+import { getVersion } from './lib/app-proxy'
 import { getOS } from '../lib/get-os'
 import { MenuEvent } from '../main-process/menu'
 import {
   Repository,
   getGitHubHtmlUrl,
-  getNonForkGitHubRepository,
   isRepositoryWithGitHubRepository,
 } from '../models/repository'
 import { Branch } from '../models/branch'
-import { PreferencesTab } from '../models/preferences'
 import { findItemByAccessKey, itemIsSelectable } from '../models/app-menu'
 import { Account } from '../models/account'
 import { TipState } from '../models/tip'
@@ -38,8 +35,6 @@ import { TitleBar, ZoomInfo, FullScreenInfo } from './window'
 
 import { RepositoriesList } from './repositories-list'
 import { RepositoryView } from './repository'
-import { RenameBranch } from './rename-branch'
-import { DeleteBranch, DeleteRemoteBranch } from './delete-branch'
 import { CloningRepositoryView } from './cloning-repository'
 import {
   Toolbar,
@@ -52,119 +47,52 @@ import {
 import { iconForRepository, OcticonSymbolType } from './octicons'
 import * as OcticonSymbol from './octicons/octicons.generated'
 import {
-  showCertificateTrustDialog,
   sendReady,
   isInApplicationFolder,
   selectAllWindowContents,
 } from './main-process-proxy'
-import { DiscardChanges } from './discard-changes'
 import { Welcome } from './welcome'
 import { AppMenuBar } from './app-menu'
 import { UpdateAvailable, renderBanner } from './banners'
-import { Preferences } from './preferences'
-import { RepositorySettings } from './repository-settings'
-import { AppError } from './app-error'
 import { MissingRepository } from './missing-repository'
-import { AddExistingRepository, CreateRepository } from './add-repository'
-import { CloneRepository } from './clone-repository'
-import { CreateBranch } from './create-branch'
-import { SignIn } from './sign-in'
-import { InstallGit } from './install-git'
-import { EditorError } from './editor'
-import { About } from './about'
-import { Publish } from './publish-repository'
-import { Acknowledgements } from './acknowledgements'
-import { UntrustedCertificate } from './untrusted-certificate'
 import { NoRepositoriesView } from './no-repositories'
-import { ConfirmRemoveRepository } from './remove-repository'
-import { TermsAndConditions } from './terms-and-conditions'
-import { PushBranchCommits } from './branches'
-import { CLIInstalled } from './cli-installed'
-import { GenericGitAuthentication } from './generic-git-auth'
-import { ShellError } from './shell'
-import { InitializeLFS, AttributeMismatch } from './lfs'
-import { UpstreamAlreadyExists } from './upstream-already-exists'
-import { ReleaseNotes } from './release-notes'
-import { DeletePullRequest } from './delete-branch/delete-pull-request-dialog'
-import { CommitConflictsWarning } from './merge-conflicts'
 import { AppTheme } from './app-theme'
 import { ApplicationTheme } from './lib/application-theme'
 import { RepositoryStateCache } from '../lib/stores/repository-state-cache'
 import { PopupType, Popup } from '../models/popup'
-import { OversizedFiles } from './changes/oversized-files-warning'
-import { PushNeedsPullWarning } from './push-needs-pull'
 import {
   ForcePushBranchState,
   getCurrentBranchForcePushState,
 } from '../lib/rebase'
 import { Banner, BannerType } from '../models/banner'
-import { StashAndSwitchBranch } from './stash-changes/stash-and-switch-branch-dialog'
-import { OverwriteStash } from './stash-changes/overwrite-stashed-changes-dialog'
-import { ConfirmDiscardStashDialog } from './stashing/confirm-discard-stash'
-import { CreateTutorialRepositoryDialog } from './no-repositories/create-tutorial-repository-dialog'
-import { ConfirmExitTutorial } from './tutorial'
 import { TutorialStep, isValidTutorialStep } from '../models/tutorial-step'
-import { WorkflowPushRejectedDialog } from './workflow-push-rejected/workflow-push-rejected'
-import { SAMLReauthRequiredDialog } from './saml-reauth-required/saml-reauth-required'
-import { CreateForkDialog } from './forks/create-fork-dialog'
 import { findContributionTargetDefaultBranch } from '../lib/branch'
-import {
-  GitHubRepository,
-  hasWritePermission,
-} from '../models/github-repository'
-import { CreateTag } from './create-tag'
-import { DeleteTag } from './delete-tag'
-import { ChooseForkSettings } from './choose-fork-settings'
-import { DiscardSelection } from './discard-changes/discard-selection-dialog'
-import { LocalChangesOverwrittenDialog } from './local-changes-overwritten/local-changes-overwritten-dialog'
-import memoizeOne from 'memoize-one'
 import { AheadBehindStore } from '../lib/stores/ahead-behind-store'
-import { getAccountForRepository } from '../lib/get-account-for-repository'
 import { CommitOneLine } from '../models/commit'
 import { CommitDragElement } from './drag-elements/commit-drag-element'
 import classNames from 'classnames'
-import { MoveToApplicationsFolder } from './move-to-applications-folder'
-import { ChangeRepositoryAlias } from './change-repository-alias/change-repository-alias-dialog'
-import { ThankYou } from './thank-you'
 import {
   getUserContributions,
   hasUserAlreadyBeenCheckedOrThanked,
   updateLastThankYou,
 } from '../lib/thank-you'
 import { ReleaseNote } from '../models/release-notes'
-import { CommitMessageDialog } from './commit-message/commit-message-dialog'
-import { buildAutocompletionProviders } from './autocompletion'
 import { DragType, DropTargetSelector } from '../models/drag-drop'
 import { dragAndDropManager } from '../lib/drag-and-drop-manager'
-import { MultiCommitOperation } from './multi-commit-operation/multi-commit-operation'
-import { WarnLocalChangesBeforeUndo } from './undo/warn-local-changes-before-undo'
-import { WarningBeforeReset } from './reset/warning-before-reset'
-import { InvalidatedToken } from './invalidated-token/invalidated-token'
 import { MultiCommitOperationKind } from '../models/multi-commit-operation'
-import { AddSSHHost } from './ssh/add-ssh-host'
-import { SSHKeyPassphrase } from './ssh/ssh-key-passphrase'
 import { getMultiCommitOperationChooseBranchStep } from '../lib/multi-commit-operation'
-import { ConfirmForcePush } from './rebase/confirm-force-push'
-import { PullRequestChecksFailed } from './notifications/pull-request-checks-failed'
-import { CICheckRunRerunDialog } from './check-runs/ci-check-run-rerun-dialog'
-import { WarnForcePushDialog } from './multi-commit-operation/dialog/warn-force-push-dialog'
 import { clamp } from '../lib/clamp'
 import { generateRepositoryListContextMenu } from './repositories-list/repository-list-item-context-menu'
 import * as ipcRenderer from '../lib/ipc-renderer'
 import { showNotification } from '../lib/notifications/show-notification'
-import { DiscardChangesRetryDialog } from './discard-changes/discard-changes-retry-dialog'
 import { generateDevReleaseSummary } from '../lib/release-notes'
-import { PullRequestReview } from './notifications/pull-request-review'
 import { getPullRequestCommitRef } from '../models/pull-request'
 import { getRepositoryType } from '../lib/git'
-import { SSHUserPassword } from './ssh/ssh-user-password'
 import { showContextualMenu } from '../lib/menu-item'
-import { UnreachableCommitsDialog } from './history/unreachable-commits-dialog'
-import { OpenPullRequestDialog } from './open-pull-request/open-pull-request-dialog'
-import { sendNonFatalException } from '../lib/helpers/non-fatal-exception'
 import { createCommitURL } from '../lib/commit-url'
 import { uuid } from '../lib/uuid'
-import { InstallingUpdate } from './installing-update/installing-update'
+import { AppPopup } from './dialog/app-popup'
+import { buildAutocompletionProviders } from './autocompletion'
 
 const MinuteInMilliseconds = 1000 * 60
 const HourInMilliseconds = MinuteInMilliseconds * 60
@@ -223,15 +151,6 @@ export class App extends React.Component<IAppProps, IAppState> {
   private get isShowingModal() {
     return this.state.currentPopup !== null
   }
-
-  /**
-   * Returns a memoized instance of onPopupDismissed() bound to the
-   * passed popupType, so it can be used in render() without creating
-   * multiple instances when the component gets re-rendered.
-   */
-  private getOnPopupDismissedFn = memoizeOne((popupId: string) => {
-    return () => this.onPopupDismissed(popupId)
-  })
 
   public constructor(props: IAppProps) {
     super(props)
@@ -1199,13 +1118,6 @@ export class App extends React.Component<IAppProps, IAppState> {
     }
   }
 
-  private onConfirmRepoRemoval = async (
-    repository: Repository,
-    deleteRepoFromDisk: boolean
-  ) => {
-    await this.props.dispatcher.removeRepository(repository, deleteRepoFromDisk)
-  }
-
   private getRepository(): Repository | CloningRepository | null {
     const state = this.state.selectedState
     if (state == null) {
@@ -1381,1097 +1293,34 @@ export class App extends React.Component<IAppProps, IAppState> {
     )
   }
 
-  private onPopupDismissed = (popupId: string) => {
-    return this.props.dispatcher.closePopupById(popupId)
-  }
-
-  private onContinueWithUntrustedCertificate = (
-    certificate: Electron.Certificate
-  ) => {
-    showCertificateTrustDialog(
-      certificate,
-      'Could not securely connect to the server, because its certificate is not trusted. Attackers might be trying to steal your information.\n\nTo connect unsafely, which may put your data at risk, you can “Always trust” the certificate and try again.'
-    )
-  }
-
   private onUpdateAvailableDismissed = () =>
     this.props.dispatcher.setUpdateBannerVisibility(false)
-
-  private currentPopupContent(): JSX.Element | null {
-    const popup = this.state.currentPopup
-
-    if (!popup) {
-      return null
-    }
-
-    if (popup.id === undefined) {
-      // Should not be possible... but if it does we want to know about it.
-      sendNonFatalException(
-        'PopupNoId',
-        new Error(
-          `Attempted to open a popup of type '${popup.type}' without an Id`
-        )
-      )
-      return null
-    }
-
-    const onPopupDismissedFn = this.getOnPopupDismissedFn(popup.id)
-
-    switch (popup.type) {
-      case PopupType.RenameBranch:
-        const stash =
-          this.state.selectedState !== null &&
-          this.state.selectedState.type === SelectionType.Repository
-            ? this.state.selectedState.state.changesState.stashEntry
-            : null
-        return (
-          <RenameBranch
-            key="rename-branch"
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            branch={popup.branch}
-            stash={stash}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.DeleteBranch:
-        return (
-          <DeleteBranch
-            key="delete-branch"
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            branch={popup.branch}
-            existsOnRemote={popup.existsOnRemote}
-            onDismissed={onPopupDismissedFn}
-            onDeleted={this.onBranchDeleted}
-          />
-        )
-      case PopupType.DeleteRemoteBranch:
-        return (
-          <DeleteRemoteBranch
-            key="delete-remote-branch"
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            branch={popup.branch}
-            onDismissed={onPopupDismissedFn}
-            onDeleted={this.onBranchDeleted}
-          />
-        )
-      case PopupType.ConfirmDiscardChanges:
-        const showSetting =
-          popup.showDiscardChangesSetting === undefined
-            ? true
-            : popup.showDiscardChangesSetting
-        const discardingAllChanges =
-          popup.discardingAllChanges === undefined
-            ? false
-            : popup.discardingAllChanges
-
-        return (
-          <DiscardChanges
-            key="discard-changes"
-            repository={popup.repository}
-            dispatcher={this.props.dispatcher}
-            files={popup.files}
-            confirmDiscardChanges={
-              this.state.askForConfirmationOnDiscardChanges
-            }
-            showDiscardChangesSetting={showSetting}
-            discardingAllChanges={discardingAllChanges}
-            onDismissed={onPopupDismissedFn}
-            onConfirmDiscardChangesChanged={this.onConfirmDiscardChangesChanged}
-          />
-        )
-      case PopupType.ConfirmDiscardSelection:
-        return (
-          <DiscardSelection
-            key="discard-selection"
-            repository={popup.repository}
-            dispatcher={this.props.dispatcher}
-            file={popup.file}
-            diff={popup.diff}
-            selection={popup.selection}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.Preferences:
-        let repository = this.getRepository()
-
-        if (repository instanceof CloningRepository) {
-          repository = null
-        }
-
-        return (
-          <Preferences
-            key="preferences"
-            initialSelectedTab={popup.initialSelectedTab}
-            dispatcher={this.props.dispatcher}
-            dotComAccount={this.getDotComAccount()}
-            confirmRepositoryRemoval={
-              this.state.askForConfirmationOnRepositoryRemoval
-            }
-            confirmDiscardChanges={
-              this.state.askForConfirmationOnDiscardChanges
-            }
-            confirmDiscardChangesPermanently={
-              this.state.askForConfirmationOnDiscardChangesPermanently
-            }
-            confirmDiscardStash={this.state.askForConfirmationOnDiscardStash}
-            confirmForcePush={this.state.askForConfirmationOnForcePush}
-            confirmUndoCommit={this.state.askForConfirmationOnUndoCommit}
-            uncommittedChangesStrategy={this.state.uncommittedChangesStrategy}
-            selectedExternalEditor={this.state.selectedExternalEditor}
-            useWindowsOpenSSH={this.state.useWindowsOpenSSH}
-            notificationsEnabled={this.state.notificationsEnabled}
-            optOutOfUsageTracking={this.state.optOutOfUsageTracking}
-            enterpriseAccount={this.getEnterpriseAccount()}
-            repository={repository}
-            onDismissed={onPopupDismissedFn}
-            selectedShell={this.state.selectedShell}
-            selectedTheme={this.state.selectedTheme}
-            customTheme={this.state.customTheme}
-            repositoryIndicatorsEnabled={this.state.repositoryIndicatorsEnabled}
-          />
-        )
-      case PopupType.RepositorySettings: {
-        const repository = popup.repository
-        const state = this.props.repositoryStateManager.get(repository)
-        const repositoryAccount = getAccountForRepository(
-          this.state.accounts,
-          repository
-        )
-
-        return (
-          <RepositorySettings
-            key={`repository-settings-${repository.hash}`}
-            initialSelectedTab={popup.initialSelectedTab}
-            remote={state.remote}
-            dispatcher={this.props.dispatcher}
-            repository={repository}
-            repositoryAccount={repositoryAccount}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.SignIn:
-        return (
-          <SignIn
-            key="sign-in"
-            signInState={this.state.signInState}
-            dispatcher={this.props.dispatcher}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.AddRepository:
-        return (
-          <AddExistingRepository
-            key="add-existing-repository"
-            onDismissed={onPopupDismissedFn}
-            dispatcher={this.props.dispatcher}
-            path={popup.path}
-          />
-        )
-      case PopupType.CreateRepository:
-        return (
-          <CreateRepository
-            key="create-repository"
-            onDismissed={onPopupDismissedFn}
-            dispatcher={this.props.dispatcher}
-            initialPath={popup.path}
-          />
-        )
-      case PopupType.CloneRepository:
-        return (
-          <CloneRepository
-            key="clone-repository"
-            dotComAccount={this.getDotComAccount()}
-            enterpriseAccount={this.getEnterpriseAccount()}
-            initialURL={popup.initialURL}
-            onDismissed={onPopupDismissedFn}
-            dispatcher={this.props.dispatcher}
-            selectedTab={this.state.selectedCloneRepositoryTab}
-            onTabSelected={this.onCloneRepositoriesTabSelected}
-            apiRepositories={this.state.apiRepositories}
-            onRefreshRepositories={this.onRefreshRepositories}
-          />
-        )
-      case PopupType.CreateBranch: {
-        const state = this.props.repositoryStateManager.get(popup.repository)
-        const branchesState = state.branchesState
-        const repository = popup.repository
-
-        if (branchesState.tip.kind === TipState.Unknown) {
-          onPopupDismissedFn()
-          return null
-        }
-
-        let upstreamGhRepo: GitHubRepository | null = null
-        let upstreamDefaultBranch: Branch | null = null
-
-        if (isRepositoryWithGitHubRepository(repository)) {
-          upstreamGhRepo = getNonForkGitHubRepository(repository)
-          upstreamDefaultBranch = branchesState.upstreamDefaultBranch
-        }
-
-        return (
-          <CreateBranch
-            key="create-branch"
-            tip={branchesState.tip}
-            defaultBranch={branchesState.defaultBranch}
-            upstreamDefaultBranch={upstreamDefaultBranch}
-            allBranches={branchesState.allBranches}
-            repository={repository}
-            targetCommit={popup.targetCommit}
-            upstreamGitHubRepository={upstreamGhRepo}
-            onBranchCreatedFromCommit={this.onBranchCreatedFromCommit}
-            onDismissed={onPopupDismissedFn}
-            dispatcher={this.props.dispatcher}
-            initialName={popup.initialName || ''}
-          />
-        )
-      }
-      case PopupType.InstallGit:
-        return (
-          <InstallGit
-            key="install-git"
-            onDismissed={onPopupDismissedFn}
-            onOpenShell={this.onOpenShellIgnoreWarning}
-            path={popup.path}
-          />
-        )
-      case PopupType.About:
-        const version = __DEV__ ? __SHA__.substring(0, 10) : getVersion()
-
-        return (
-          <About
-            key="about"
-            onDismissed={onPopupDismissedFn}
-            applicationName={getName()}
-            applicationVersion={version}
-            applicationArchitecture={process.arch}
-            onCheckForUpdates={this.onCheckForUpdates}
-            onCheckForNonStaggeredUpdates={this.onCheckForNonStaggeredUpdates}
-            onShowAcknowledgements={this.showAcknowledgements}
-            onShowTermsAndConditions={this.showTermsAndConditions}
-          />
-        )
-      case PopupType.PublishRepository:
-        return (
-          <Publish
-            key="publish"
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            accounts={this.state.accounts}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.UntrustedCertificate:
-        return (
-          <UntrustedCertificate
-            key="untrusted-certificate"
-            certificate={popup.certificate}
-            url={popup.url}
-            onDismissed={onPopupDismissedFn}
-            onContinue={this.onContinueWithUntrustedCertificate}
-          />
-        )
-      case PopupType.Acknowledgements:
-        return (
-          <Acknowledgements
-            key="acknowledgements"
-            onDismissed={onPopupDismissedFn}
-            applicationVersion={getVersion()}
-          />
-        )
-      case PopupType.RemoveRepository:
-        return (
-          <ConfirmRemoveRepository
-            key="confirm-remove-repository"
-            repository={popup.repository}
-            onConfirmation={this.onConfirmRepoRemoval}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.TermsAndConditions:
-        return (
-          <TermsAndConditions
-            key="terms-and-conditions"
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.PushBranchCommits:
-        return (
-          <PushBranchCommits
-            key="push-branch-commits"
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            branch={popup.branch}
-            unPushedCommits={popup.unPushedCommits}
-            onConfirm={this.openCreatePullRequestInBrowser}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.CLIInstalled:
-        return (
-          <CLIInstalled key="cli-installed" onDismissed={onPopupDismissedFn} />
-        )
-      case PopupType.GenericGitAuthentication:
-        return (
-          <GenericGitAuthentication
-            key="generic-git-authentication"
-            hostname={popup.hostname}
-            onDismiss={onPopupDismissedFn}
-            onSave={this.onSaveCredentials}
-            retryAction={popup.retryAction}
-          />
-        )
-      case PopupType.ExternalEditorFailed:
-        const openPreferences = popup.openPreferences
-        const suggestDefaultEditor = popup.suggestDefaultEditor
-
-        return (
-          <EditorError
-            key="editor-error"
-            message={popup.message}
-            onDismissed={onPopupDismissedFn}
-            showPreferencesDialog={this.onShowAdvancedPreferences}
-            viewPreferences={openPreferences}
-            suggestDefaultEditor={suggestDefaultEditor}
-          />
-        )
-      case PopupType.OpenShellFailed:
-        return (
-          <ShellError
-            key="shell-error"
-            message={popup.message}
-            onDismissed={onPopupDismissedFn}
-            showPreferencesDialog={this.onShowAdvancedPreferences}
-          />
-        )
-      case PopupType.InitializeLFS:
-        return (
-          <InitializeLFS
-            key="initialize-lfs"
-            repositories={popup.repositories}
-            onDismissed={onPopupDismissedFn}
-            onInitialize={this.initializeLFS}
-          />
-        )
-      case PopupType.LFSAttributeMismatch:
-        return (
-          <AttributeMismatch
-            key="lsf-attribute-mismatch"
-            onDismissed={onPopupDismissedFn}
-            onUpdateExistingFilters={this.updateExistingLFSFilters}
-          />
-        )
-      case PopupType.UpstreamAlreadyExists:
-        return (
-          <UpstreamAlreadyExists
-            key="upstream-already-exists"
-            repository={popup.repository}
-            existingRemote={popup.existingRemote}
-            onDismissed={onPopupDismissedFn}
-            onUpdate={this.onUpdateExistingUpstreamRemote}
-            onIgnore={this.onIgnoreExistingUpstreamRemote}
-          />
-        )
-      case PopupType.ReleaseNotes:
-        return (
-          <ReleaseNotes
-            key="release-notes"
-            emoji={this.state.emoji}
-            newReleases={popup.newReleases}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.DeletePullRequest:
-        return (
-          <DeletePullRequest
-            key="delete-pull-request"
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            branch={popup.branch}
-            onDismissed={onPopupDismissedFn}
-            pullRequest={popup.pullRequest}
-          />
-        )
-      case PopupType.OversizedFiles:
-        return (
-          <OversizedFiles
-            key="oversized-files"
-            oversizedFiles={popup.oversizedFiles}
-            onDismissed={onPopupDismissedFn}
-            dispatcher={this.props.dispatcher}
-            context={popup.context}
-            repository={popup.repository}
-          />
-        )
-      case PopupType.CommitConflictsWarning:
-        return (
-          <CommitConflictsWarning
-            key="commit-conflicts-warning"
-            dispatcher={this.props.dispatcher}
-            files={popup.files}
-            repository={popup.repository}
-            context={popup.context}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.PushNeedsPull:
-        return (
-          <PushNeedsPullWarning
-            key="push-needs-pull"
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.ConfirmForcePush: {
-        const { askForConfirmationOnForcePush } = this.state
-
-        return (
-          <ConfirmForcePush
-            key="confirm-force-push"
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            upstreamBranch={popup.upstreamBranch}
-            askForConfirmationOnForcePush={askForConfirmationOnForcePush}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.StashAndSwitchBranch: {
-        const { repository, branchToCheckout } = popup
-        const { branchesState, changesState } =
-          this.props.repositoryStateManager.get(repository)
-        const { tip } = branchesState
-
-        if (tip.kind !== TipState.Valid) {
-          return null
-        }
-
-        const currentBranch = tip.branch
-        const hasAssociatedStash = changesState.stashEntry !== null
-
-        return (
-          <StashAndSwitchBranch
-            key="stash-and-switch-branch"
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            currentBranch={currentBranch}
-            branchToCheckout={branchToCheckout}
-            hasAssociatedStash={hasAssociatedStash}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.ConfirmOverwriteStash: {
-        const { repository, branchToCheckout: branchToCheckout } = popup
-        return (
-          <OverwriteStash
-            key="overwrite-stash"
-            dispatcher={this.props.dispatcher}
-            repository={repository}
-            branchToCheckout={branchToCheckout}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.ConfirmDiscardStash: {
-        const { repository, stash } = popup
-
-        return (
-          <ConfirmDiscardStashDialog
-            key="confirm-discard-stash-dialog"
-            dispatcher={this.props.dispatcher}
-            askForConfirmationOnDiscardStash={
-              this.state.askForConfirmationOnDiscardStash
-            }
-            repository={repository}
-            stash={stash}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.CreateTutorialRepository: {
-        return (
-          <CreateTutorialRepositoryDialog
-            key="create-tutorial-repository-dialog"
-            account={popup.account}
-            progress={popup.progress}
-            onDismissed={onPopupDismissedFn}
-            onCreateTutorialRepository={this.onCreateTutorialRepository}
-          />
-        )
-      }
-      case PopupType.ConfirmExitTutorial: {
-        return (
-          <ConfirmExitTutorial
-            key="confirm-exit-tutorial"
-            onDismissed={onPopupDismissedFn}
-            onContinue={this.onExitTutorialToHomeScreen}
-          />
-        )
-      }
-      case PopupType.PushRejectedDueToMissingWorkflowScope:
-        return (
-          <WorkflowPushRejectedDialog
-            onDismissed={onPopupDismissedFn}
-            rejectedPath={popup.rejectedPath}
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-          />
-        )
-      case PopupType.SAMLReauthRequired:
-        return (
-          <SAMLReauthRequiredDialog
-            onDismissed={onPopupDismissedFn}
-            organizationName={popup.organizationName}
-            endpoint={popup.endpoint}
-            retryAction={popup.retryAction}
-            dispatcher={this.props.dispatcher}
-          />
-        )
-      case PopupType.CreateFork:
-        return (
-          <CreateForkDialog
-            onDismissed={onPopupDismissedFn}
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            account={popup.account}
-          />
-        )
-      case PopupType.CreateTag: {
-        return (
-          <CreateTag
-            key="create-tag"
-            repository={popup.repository}
-            onDismissed={onPopupDismissedFn}
-            dispatcher={this.props.dispatcher}
-            targetCommitSha={popup.targetCommitSha}
-            initialName={popup.initialName}
-            localTags={popup.localTags}
-          />
-        )
-      }
-      case PopupType.DeleteTag: {
-        return (
-          <DeleteTag
-            key="delete-tag"
-            repository={popup.repository}
-            onDismissed={onPopupDismissedFn}
-            dispatcher={this.props.dispatcher}
-            tagName={popup.tagName}
-          />
-        )
-      }
-      case PopupType.ChooseForkSettings: {
-        return (
-          <ChooseForkSettings
-            repository={popup.repository}
-            onDismissed={onPopupDismissedFn}
-            dispatcher={this.props.dispatcher}
-          />
-        )
-      }
-      case PopupType.LocalChangesOverwritten:
-        const selectedState = this.state.selectedState
-
-        const existingStash =
-          selectedState !== null &&
-          selectedState.type === SelectionType.Repository
-            ? selectedState.state.changesState.stashEntry
-            : null
-
-        return (
-          <LocalChangesOverwrittenDialog
-            repository={popup.repository}
-            dispatcher={this.props.dispatcher}
-            hasExistingStash={existingStash !== null}
-            retryAction={popup.retryAction}
-            onDismissed={onPopupDismissedFn}
-            files={popup.files}
-          />
-        )
-      case PopupType.MoveToApplicationsFolder: {
-        return (
-          <MoveToApplicationsFolder
-            dispatcher={this.props.dispatcher}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.ChangeRepositoryAlias: {
-        return (
-          <ChangeRepositoryAlias
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.ThankYou:
-        return (
-          <ThankYou
-            key="thank-you"
-            emoji={this.state.emoji}
-            userContributions={popup.userContributions}
-            friendlyName={popup.friendlyName}
-            latestVersion={popup.latestVersion}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      case PopupType.CommitMessage:
-        const repositoryState = this.props.repositoryStateManager.get(
-          popup.repository
-        )
-
-        const { tip } = repositoryState.branchesState
-        const currentBranchName: string | null =
-          tip.kind === TipState.Valid ? tip.branch.name : null
-
-        const hasWritePermissionForRepository =
-          popup.repository.gitHubRepository === null ||
-          hasWritePermission(popup.repository.gitHubRepository)
-
-        const autocompletionProviders = buildAutocompletionProviders(
-          popup.repository,
-          this.props.dispatcher,
-          this.state.emoji,
-          this.props.issuesStore,
-          this.props.gitHubUserStore,
-          this.state.accounts
-        )
-
-        const repositoryAccount = getAccountForRepository(
-          this.state.accounts,
-          popup.repository
-        )
-
-        return (
-          <CommitMessageDialog
-            key="commit-message"
-            autocompletionProviders={autocompletionProviders}
-            branch={currentBranchName}
-            coAuthors={popup.coAuthors}
-            commitAuthor={repositoryState.commitAuthor}
-            commitMessage={popup.commitMessage}
-            commitSpellcheckEnabled={this.state.commitSpellcheckEnabled}
-            dialogButtonText={popup.dialogButtonText}
-            dialogTitle={popup.dialogTitle}
-            dispatcher={this.props.dispatcher}
-            prepopulateCommitSummary={popup.prepopulateCommitSummary}
-            repository={popup.repository}
-            showBranchProtected={
-              repositoryState.changesState.currentBranchProtected
-            }
-            showCoAuthoredBy={popup.showCoAuthoredBy}
-            showNoWriteAccess={!hasWritePermissionForRepository}
-            onDismissed={onPopupDismissedFn}
-            onSubmitCommitMessage={popup.onSubmitCommitMessage}
-            repositoryAccount={repositoryAccount}
-          />
-        )
-      case PopupType.MultiCommitOperation: {
-        const { selectedState, emoji } = this.state
-
-        if (
-          selectedState === null ||
-          selectedState.type !== SelectionType.Repository
-        ) {
-          return null
-        }
-
-        const { changesState, multiCommitOperationState } = selectedState.state
-        const { workingDirectory, conflictState } = changesState
-        if (multiCommitOperationState === null) {
-          log.warn(
-            '[App] invalid state encountered - multi commit flow should not be active when step is null'
-          )
-          return null
-        }
-
-        return (
-          <MultiCommitOperation
-            key="multi-commit-operation"
-            repository={popup.repository}
-            dispatcher={this.props.dispatcher}
-            state={multiCommitOperationState}
-            conflictState={conflictState}
-            emoji={emoji}
-            workingDirectory={workingDirectory}
-            askForConfirmationOnForcePush={
-              this.state.askForConfirmationOnForcePush
-            }
-            openFileInExternalEditor={this.openFileInExternalEditor}
-            resolvedExternalEditor={this.state.resolvedExternalEditor}
-            openRepositoryInShell={this.openCurrentRepositoryInShell}
-          />
-        )
-      }
-      case PopupType.WarnLocalChangesBeforeUndo: {
-        const { repository, commit, isWorkingDirectoryClean } = popup
-        return (
-          <WarnLocalChangesBeforeUndo
-            key="warn-local-changes-before-undo"
-            dispatcher={this.props.dispatcher}
-            repository={repository}
-            commit={commit}
-            isWorkingDirectoryClean={isWorkingDirectoryClean}
-            confirmUndoCommit={this.state.askForConfirmationOnUndoCommit}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.WarningBeforeReset: {
-        const { repository, commit } = popup
-        return (
-          <WarningBeforeReset
-            key="warning-before-reset"
-            dispatcher={this.props.dispatcher}
-            repository={repository}
-            commit={commit}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.InvalidatedToken: {
-        return (
-          <InvalidatedToken
-            key="invalidated-token"
-            dispatcher={this.props.dispatcher}
-            account={popup.account}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.AddSSHHost: {
-        return (
-          <AddSSHHost
-            key="add-ssh-host"
-            host={popup.host}
-            ip={popup.ip}
-            keyType={popup.keyType}
-            fingerprint={popup.fingerprint}
-            onSubmit={popup.onSubmit}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.SSHKeyPassphrase: {
-        return (
-          <SSHKeyPassphrase
-            key="ssh-key-passphrase"
-            keyPath={popup.keyPath}
-            onSubmit={popup.onSubmit}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.SSHUserPassword: {
-        return (
-          <SSHUserPassword
-            key="ssh-user-password"
-            username={popup.username}
-            onSubmit={popup.onSubmit}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.PullRequestChecksFailed: {
-        return (
-          <PullRequestChecksFailed
-            key="pull-request-checks-failed"
-            dispatcher={this.props.dispatcher}
-            shouldChangeRepository={popup.shouldChangeRepository}
-            repository={popup.repository}
-            pullRequest={popup.pullRequest}
-            commitMessage={popup.commitMessage}
-            commitSha={popup.commitSha}
-            checks={popup.checks}
-            accounts={this.state.accounts}
-            onSubmit={onPopupDismissedFn}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.CICheckRunRerun: {
-        return (
-          <CICheckRunRerunDialog
-            key="rerun-check-runs"
-            checkRuns={popup.checkRuns}
-            dispatcher={this.props.dispatcher}
-            repository={popup.repository}
-            prRef={popup.prRef}
-            onDismissed={onPopupDismissedFn}
-            failedOnly={popup.failedOnly}
-          />
-        )
-      }
-      case PopupType.WarnForcePush: {
-        const { askForConfirmationOnForcePush } = this.state
-        return (
-          <WarnForcePushDialog
-            key="warn-force-push"
-            dispatcher={this.props.dispatcher}
-            operation={popup.operation}
-            askForConfirmationOnForcePush={askForConfirmationOnForcePush}
-            onBegin={this.getWarnForcePushDialogOnBegin(
-              popup.onBegin,
-              onPopupDismissedFn
-            )}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.DiscardChangesRetry: {
-        return (
-          <DiscardChangesRetryDialog
-            key="discard-changes-retry"
-            dispatcher={this.props.dispatcher}
-            retryAction={popup.retryAction}
-            onDismissed={onPopupDismissedFn}
-            onConfirmDiscardChangesChanged={
-              this.onConfirmDiscardChangesPermanentlyChanged
-            }
-          />
-        )
-      }
-      case PopupType.PullRequestReview: {
-        return (
-          <PullRequestReview
-            key="pull-request-checks-failed"
-            dispatcher={this.props.dispatcher}
-            shouldCheckoutBranch={popup.shouldCheckoutBranch}
-            shouldChangeRepository={popup.shouldChangeRepository}
-            repository={popup.repository}
-            pullRequest={popup.pullRequest}
-            review={popup.review}
-            numberOfComments={popup.numberOfComments}
-            emoji={this.state.emoji}
-            accounts={this.state.accounts}
-            onSubmit={onPopupDismissedFn}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.UnreachableCommits: {
-        const { selectedState, emoji } = this.state
-        if (
-          selectedState == null ||
-          selectedState.type !== SelectionType.Repository
-        ) {
-          return null
-        }
-
-        const {
-          commitLookup,
-          commitSelection: { shas, shasInDiff },
-        } = selectedState.state
-
-        return (
-          <UnreachableCommitsDialog
-            selectedShas={shas}
-            shasInDiff={shasInDiff}
-            commitLookup={commitLookup}
-            selectedTab={popup.selectedTab}
-            emoji={emoji}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.StartPullRequest: {
-        // Intentionally chose to get the current pull request state  on
-        // rerender because state variables such as file selection change
-        // via the dispatcher.
-        const pullRequestState = this.getPullRequestState()
-        if (pullRequestState === null) {
-          // This shouldn't happen..
-          sendNonFatalException(
-            'FailedToStartPullRequest',
-            new Error(
-              'Failed to start pull request because pull request state was null'
-            )
-          )
-          return null
-        }
-
-        const { pullRequestFilesListWidth, hideWhitespaceInPullRequestDiff } =
-          this.state
-
-        const {
-          allBranches,
-          currentBranch,
-          defaultBranch,
-          imageDiffType,
-          externalEditorLabel,
-          nonLocalCommitSHA,
-          recentBranches,
-          repository,
-          showSideBySideDiff,
-        } = popup
-
-        return (
-          <OpenPullRequestDialog
-            key="open-pull-request"
-            allBranches={allBranches}
-            currentBranch={currentBranch}
-            defaultBranch={defaultBranch}
-            dispatcher={this.props.dispatcher}
-            fileListWidth={pullRequestFilesListWidth}
-            hideWhitespaceInDiff={hideWhitespaceInPullRequestDiff}
-            imageDiffType={imageDiffType}
-            nonLocalCommitSHA={nonLocalCommitSHA}
-            pullRequestState={pullRequestState}
-            recentBranches={recentBranches}
-            repository={repository}
-            externalEditorLabel={externalEditorLabel}
-            showSideBySideDiff={showSideBySideDiff}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      case PopupType.Error: {
-        return (
-          <AppError
-            error={popup.error}
-            onDismissed={onPopupDismissedFn}
-            onShowPopup={this.showPopup}
-            onRetryAction={this.onRetryAction}
-          />
-        )
-      }
-      case PopupType.InstallingUpdate: {
-        return (
-          <InstallingUpdate
-            key="installing-update"
-            dispatcher={this.props.dispatcher}
-            onDismissed={onPopupDismissedFn}
-          />
-        )
-      }
-      default:
-        return assertNever(popup, `Unknown popup type: ${popup}`)
-    }
-  }
-
-  private getPullRequestState() {
-    const { selectedState } = this.state
-    if (
-      selectedState == null ||
-      selectedState.type !== SelectionType.Repository
-    ) {
-      return null
-    }
-
-    return selectedState.state.pullRequestState
-  }
-
-  private getWarnForcePushDialogOnBegin(
-    onBegin: () => void,
-    onPopupDismissedFn: () => void
-  ) {
-    return () => {
-      onBegin()
-      onPopupDismissedFn()
-    }
-  }
-
-  private onExitTutorialToHomeScreen = () => {
-    const tutorialRepository = this.getSelectedTutorialRepository()
-    if (!tutorialRepository) {
-      return false
-    }
-
-    this.props.dispatcher.pauseTutorial(tutorialRepository)
-    return true
-  }
-
-  private onCreateTutorialRepository = (account: Account) => {
-    this.props.dispatcher.createTutorialRepository(account)
-  }
-
-  private onUpdateExistingUpstreamRemote = (repository: Repository) => {
-    this.props.dispatcher.updateExistingUpstreamRemote(repository)
-  }
-
-  private onIgnoreExistingUpstreamRemote = (repository: Repository) => {
-    this.props.dispatcher.ignoreExistingUpstreamRemote(repository)
-  }
-
-  private updateExistingLFSFilters = () => {
-    this.props.dispatcher.installGlobalLFSFilters(true)
-  }
-
-  private initializeLFS = (repositories: ReadonlyArray<Repository>) => {
-    this.props.dispatcher.installLFSHooks(repositories)
-  }
-
-  private onCloneRepositoriesTabSelected = (tab: CloneRepositoryTab) => {
-    this.props.dispatcher.changeCloneRepositoriesTab(tab)
-  }
 
   private onRefreshRepositories = (account: Account) => {
     this.props.dispatcher.refreshApiRepositories(account)
   }
 
-  private onShowAdvancedPreferences = () => {
-    this.props.dispatcher.showPopup({
-      type: PopupType.Preferences,
-      initialSelectedTab: PreferencesTab.Advanced,
-    })
-  }
-
-  private onBranchCreatedFromCommit = () => {
-    const repositoryView = this.repositoryViewRef.current
-    if (repositoryView !== null) {
-      repositoryView.scrollCompareListToTop()
-    }
-  }
-
-  private onOpenShellIgnoreWarning = (path: string) => {
-    this.props.dispatcher.openShell(path, true)
-  }
-
-  private onSaveCredentials = async (
-    hostname: string,
-    username: string,
-    password: string,
-    retryAction: RetryAction
-  ) => {
-    await this.props.dispatcher.saveGenericGitCredentials(
-      hostname,
-      username,
-      password
+  private buildAutocompletionProviders = (repository: Repository) => {
+    return buildAutocompletionProviders(
+      repository,
+      this.props.dispatcher,
+      this.state.emoji,
+      this.props.issuesStore,
+      this.props.gitHubUserStore,
+      this.state.accounts
     )
-
-    this.props.dispatcher.performRetry(retryAction)
-  }
-
-  private onCheckForUpdates = () => this.checkForUpdates(false)
-  private onCheckForNonStaggeredUpdates = () =>
-    this.checkForUpdates(false, true)
-
-  private showAcknowledgements = () => {
-    this.props.dispatcher.showPopup({ type: PopupType.Acknowledgements })
-  }
-
-  private showTermsAndConditions = () => {
-    this.props.dispatcher.showPopup({ type: PopupType.TermsAndConditions })
   }
 
   private renderPopup() {
-    const popupContent = this.currentPopupContent()
-
     return (
-      <TransitionGroup>
-        {popupContent && (
-          <CSSTransition classNames="modal" timeout={dialogTransitionTimeout}>
-            {popupContent}
-          </CSSTransition>
-        )}
-      </TransitionGroup>
+      <AppPopup
+        appState={this.state}
+        dispatcher={this.props.dispatcher}
+        repositoryViewRef={this.repositoryViewRef}
+        repositoryStateManager={this.props.repositoryStateManager}
+        checkForUpdates={this.checkForUpdates}
+        buildAutocompletionProviders={this.buildAutocompletionProviders}
+      />
     )
   }
 
@@ -2514,18 +1363,6 @@ export class App extends React.Component<IAppProps, IAppState> {
 
   private renderFullScreenInfo() {
     return <FullScreenInfo windowState={this.state.windowState} />
-  }
-
-  private onConfirmDiscardChangesChanged = (value: boolean) => {
-    this.props.dispatcher.setConfirmDiscardChangesSetting(value)
-  }
-
-  private onConfirmDiscardChangesPermanentlyChanged = (value: boolean) => {
-    this.props.dispatcher.setConfirmDiscardChangesPermanentlySetting(value)
-  }
-
-  private onRetryAction = (retryAction: RetryAction) => {
-    this.props.dispatcher.performRetry(retryAction)
   }
 
   private showPopup = (popup: Popup) => {
@@ -2851,13 +1688,6 @@ export class App extends React.Component<IAppProps, IAppState> {
     this.props.dispatcher.startPullRequest(state.repository)
   }
 
-  private openCreatePullRequestInBrowser = (
-    repository: Repository,
-    branch: Branch
-  ) => {
-    this.props.dispatcher.openCreatePullRequestInBrowser(repository, branch)
-  }
-
   private onBranchDropdownStateChanged = (newState: DropdownState) => {
     if (newState === 'open') {
       this.props.dispatcher.showFoldout({ type: FoldoutType.Branch })
@@ -3138,16 +1968,6 @@ export class App extends React.Component<IAppProps, IAppState> {
     }
 
     this.props.dispatcher.openInBrowser(commitURL)
-  }
-
-  private onBranchDeleted = (repository: Repository) => {
-    // In the event a user is in the middle of a compare
-    // we need to exit out of the compare state after the
-    // branch has been deleted. Calling executeCompare allows
-    // us to do just that.
-    this.props.dispatcher.executeCompare(repository, {
-      kind: HistoryTabMode.History,
-    })
   }
 
   private inNoRepositoriesViewState() {
